seems like episode detail page works
