+++ dashboard/app/routes/observability/episode/EpisodeFeedbackTable.tsx
@@ -6,11 +6,12 @@ import {
   TableHeader,
   TableRow,
 } from "~/components/ui/table";
-import { getMetricName, type FeedbackRow } from "~/utils/clickhouse/feedback";
+import FeedbackValue from "~/components/feedback/FeedbackValue";
+import { getMetricName } from "~/utils/clickhouse/helpers";
+import type { FeedbackRow } from "~/utils/clickhouse/feedback";
 import { formatDate } from "~/utils/date";
 import { MetricBadges } from "~/components/metric/MetricBadges";
 import { useConfig } from "~/context/config";
-import FeedbackValue from "~/components/feedback/FeedbackValue";
 
 export default function EpisodeFeedbackTable({
   feedback,
