+++ dashboard/app/utils/clickhouse/helpers.ts
@@ -1,0 +1,13 @@
+import type { FeedbackRow } from "./feedback";
+
+// Since demonstrations and comments do not have a metric_name, we need to
+// infer the metric name from the structure of the feedback row
+export const getMetricName = (feedback: FeedbackRow) => {
+  if ("metric_name" in feedback) {
+    return feedback.metric_name;
+  }
+  if ("inference_id" in feedback) {
+    return "demonstration";
+  }
+  return "comment";
+};
